Implement more AST classes to make the transpiler work for emptyclass.
This CL adds/modifies some needed AST structures and velocity template files to make the transpiler be able to transpile an emptyclass. While some detailed implementations are still missing. (For example, computing mangled name for different types of classes, and so on).

Currently the import part (goog.require(...)) is still not correct. I think we need a pass to collect the imports information after the entire AST is built (we cannot do it during AST building due to the lack of information, for example, the compilationUnit name of the type reference), which will depend on our ast visitor.

PiperOrigin-RevId: 95786123