Add AST node and readability test for variable declaration and new instance.
Also move code generation for Statement to a central place (StatementSourceGenerator). Previously it is implemented by toString() method in each AST node. But it will rely on some utility functions defined in TranspilerUtils and ManglingNameUtils, which should be placed in package generator/. We should implement the code generation logic in generator/ but not in ast/. So I think the change makes sense.

PiperOrigin-RevId: 96353334