Add for statement
This also includes some refactoring of the AST to properly represent variable declaration expressions and statements.

Before we would always model declaring a variable as a statement, this is not true for a for loop. In JDT initializers in a for loop are expressions.

Example:
for (int i = 0; j = 1;;)

  ForStatement:
    initializers:
      VariableDeclarationExpression (type goes here)
        VariableDeclarationFragment
          -> variable name, initializer

int j = 0;
  VariableDeclarationStatement: (type goes here)
    VariableDeclarationFragment
	  -> variable name; initializer

PiperOrigin-RevId: 97572778