Add support for instance inner class.
The access to the enclosing instances are translated in CompilationUnitBuilder. The constructors, and calls to constructors (NewInstance, super() call and this() call) are normalized by NormalizeNestedClassConstructorsVisitor.

If a child class and its parent class share some enclosing instances, an implicit super() call is allowed. In this case, we cannot synthesize super() call outside CompilationUnitBuilder due to the limited information in our AST. So moved the createDefaultConstructorsVisitor and InsertExplicitSuperCallsVisitor to CompilationUnitBuilder.

PiperOrigin-RevId: 98237494