Works around Closure's inconsistent opinion of the type of classes.
When referencing class and interface definitions from outside of their module
Closure thinks of them as 'Function' but when referencing them from inside
their module Closure thinks of them as 'function (this:foo.bar.Type)'.

This inconsistency is already tracked in a bug (b/20102666) but it would be
nice to not accumulate another warning everytime we create a new interface
class and that interface class definition marks itself with an initial
Foo.$markImplementor(Foo); call.

So I have added a cast so that Closure will think of type definitions as
'Function' even in the original defining module.

PiperOrigin-RevId: 98663115