Switches Long temporary vars from window.$ to Longs.$.
Previously Closure compiler crashed when accessing Longs.$
temporary vars inside of a comma expression.

But they have fixed the crash and so we can switch over to the
more readable version. Also it turns out that it compresses
slightly better because Closure compiler can minify Longs.$q
and Longs.$v each to a single 1 letter var, which it wasn't
able to do with the window.$q and window.$v references.

PiperOrigin-RevId: 98847025