support for simple lambdas.
Lambda expression is transpiled to an inner class:

class Enclosing$lambda$0 implements I {
  Enclosing$lambda$0(captures) {// initialize captures}
  private T lambda$0(args) {...lambda_expression_with_captures ... }
  public T samMethod0(args) { return this.lambda$0(args); }
}
And replaces the lambda with new Enclosing$lambda$0(captures).

lambda function is created in the lambda inner class. In this way, it works more consistently as an inner class, and a little more readable than creating in enclosing class.

PiperOrigin-RevId: 101261115