Adds --remove_unused_constructor_properties=ON, trims some @nocollapse size.
Ben mentioned --remove_unused_constructor_properties=ON as a way to address
some of the unused static methods that are getting left behind after having
to temporarily turn off collapse_properties. This is a nice stopgap (but
also Ben is still working on improved goog.module.get() handling that will
enable us to remove a lot of the @nocollapse annotations).

PiperOrigin-RevId: 101270584