Adds optional Zip bundle output.
With this option instead of outputting multiple JS files in a
directory the generated JS files will end up in a single bundle zip.

Since blaze is not yet ready to process JS zip bundles (and won't be
ready for at least another week even in canary) this option is not
yet being used by the .bzl files but an updated j2cl_transpile_zip
target is provided to start testing once the feature is available in
blaze canary.

PiperOrigin-RevId: 101586874