Fixed a lambda issue.
A lambda body that is an Expression was always translated to a ReturnStatement, which is not correct if the corresponding SAM method does not return a value. It should be translated to a ReturnStatement (return type of SAM method is not void) or an ExpressionStatement (return type of SAM method is void).

PiperOrigin-RevId: 103129198