fix calls to outer class's function.
especially for the cases where a function is called by an implicit 'this' reference, which can refers to any enclosing instances, or itself, or the super instance.

Also fix bugs in local class/anonymous class that is declared in a static context. A local class that is declared in a static context should not have an enclosing instance. (although the local class itself is not static).

PiperOrigin-RevId: 103424445