Fixes a bug when gathering imports for ArrayTypeDescriptors.
The arrayTypeDescriptor.getLeafTypeDescriptor() value was being
directly added to the import list. When the leaf type was generic
it resuled in the addition of bogus imports.

The fix is to pass the leafTypeDescriptor back through the same
add() function that already knows how to filter out type
parameters and unwrap generic types.

PiperOrigin-RevId: 103664224