Add JUnit3 support & improve JUnit4 support
Almost all common testing patterns are now supported. This CL only does Unit testing, integration testing still needs to be done.

The last missing feature is to run tests that are inner classes, for now our code generation does not bother with these and just throws an exception. This should be added laster.

PiperOrigin-RevId: 103980138