Fix a bug on casting in a bridge method.
Erasure type was used as the cast-to type to make sure casting to a type variable was translated without warnings. However it leads to a bug if the cast-to type is a parameterized type, whose erasure type is the generic type. JsDoc type annotation for the generic type will introduce unknown type variables.

Now that cast to generics is fixed in NormalizeCastVisitors, the real accurate type should be used when creating a bridge method.

PiperOrigin-RevId: 104326166