Cleanup MethodDescriptor, and fixed a bug in bridge method.
Removed erasureMethodDescriptor field in MethodDescriptor. This field causes some confusion and actually it is not useful anymore. Previously it was used to generate cast expression for the argument of a generic method call. Later we decided not to generate the cast expression. So actually we can get rid of erasureMethodDescriptor. Now whenever we create a MethodDescriptor, we use the rawTypeDescriptor for parameter types.

Previously redundant bridge methods were created if two methods have the same parameter signature but different return types. It is solved by setting the return type of the bridge method to the return type of the delegated method. It is also more accurate.

PiperOrigin-RevId: 104435986