workaround for unresolved template declared by a method.
See b/24476009. It is a bug in JSCompiler's old type checker. Template declared in a method is not resolved inside the method. They're not going to fix it in the old type checker. Switching to new type inference causes many warnings. Although we have to switch to the new type inference someday, it may be easier when their new type checker is more stable and produces fewer false positives and addresses more existing issues.

This workaround replaces the template in a method with its bound. Although it is not one hundred percent accurate, I think it is reasonable.

PiperOrigin-RevId: 104510099