Fixes a parentheses bug in unboxing of ternary expressions.
Both JDT and J2CL have a ParenthesizedExpression node and you can see
it in the tree and we emit in the output JS. You can see it when
representing code that started as Java that looks like this:
d = (a ? b : c).run();
because the user wrote the parentheses themselves.

But sometimes the user can omit those parenthesis like this:
d = a ? b : c;

But if autoboxing needed to insert some unboxing method call on the
result of the ternary expression we were failing to insert the
parentheses that the user would (under normal circumstances) have
written. Ending up with emitted JS that looked liked:
d = a ? b : c.intValue();
when it should have looked like:
d = (a ? b : c).intValue();
PiperOrigin-RevId: 104678531