Cleanup in CompilationUnitBuilder and NormalizeNestedClassConstructorsVisitor.
1. Moved synthesizeDefaultConstructor and synthesizeSuperCall out of CompilationUnitBuilder to separate passes.
2. Moved some nested class normalizations to NormalizeNestedClassConstructorsVisitor.

Now the line between CompilationUnitBuilder and NormalizeNestedClassConstructorsVisitor is much cleaner. The qualifiers and variables are resolved in CompilationUnitBuilder. And the normalizations and structural AST changes are done in NormalizeNestedClassConstructorsVisitor.

Follow up cleanup: current NormalizationNestedClassConstructorsVisitor does some normalizations following some implicit orders, for example, rewriteFiedAccess() which fixes some fieldAccess and rewriteMethod() which adds field initializers are done in one pass, it is not obvious if the fiedAccess in added initializers are fixed. It should be better to have them done in separate passes.

PiperOrigin-RevId: 105015839