Fixed the remaining bug in anonymous class creation.
Create an anonymous class of an inner class is not transpiled correctly. For example,
A a = q.new A() {}; A is an inner class. the qualifier q is actually the enclosing instance of the super instance. This is transpiled like this:

class $1 extends A {
  $1($sq) {$sq.super();} // Add an extra parameter to the anonymous class's constructor, and use it as the qualifier of the super call.
}

a = new $1(q); // pass the qualifier of the new statement as an argument.
PiperOrigin-RevId: 105323573