support for instance js methods.
A bridge method is generated for direct JsMethod that exposes a non-JsMethod (overrides a non-JsMethod, and is the first at the overriding chain). The bridge method has regularly mangled name and delegates to the unmangled js method. Overriding methods of js methods have the same unmangled name as js method. Method calls to JsMethod or its overriding methods are transpiled to method calls to unmangled method.

PiperOrigin-RevId: 107490879