Stopped abstract types from replacing already implemented methods.
Abstract types are implemented in a way to output a stub function
for every not-yet-implemented method that is required by any
interfaces they implement.

The list of stub functions was not supposed to override methods
already provided by parent classes (which in combination with late
interfaces are known as "accidental overrides") but that
calculation was not being done quite right. As a result some
already implemented parent class methods were being replaced.

PiperOrigin-RevId: 107546141