This CL is a temporary workaround for https://buganizer.corp.google.com/u/0/issues/25512693
JSCompiler is incorrectly copying the value of getter methods when sub classing which invokes the getter during the declaration phase and causes the clinit to be called too quickly.

In this CL we move the clinit calls to the call site to prevent this behavior and remove getters and setters completely.

We add the clinit calls to the top of method bodys that contain static field access's.  We also add clinit calls to constructors when instance field initializers access static fields of another class.  For a class A that has a static field initializer that accesses a static field on class B, we add B.clinit() inside A.clinit right before we initialize A's static fields.

Check out readable/callsiteclinit for examples.

PiperOrigin-RevId: 107609415