Fix missing bridge methods issue.
Bridge method was generated only for generic method that has a concrete overriding method. The bridge method contains a forwarding call to a concrete method. We're missing the case when a concrete interface method is overridden by a generic method. In this case a backward bridge method (from the concrete method to a generic method) should be generated.

PiperOrigin-RevId: 107719660