Trims unreferenced file fluff out of JSCompiler output.
Uses --only_closure_dependencies and closure_entry_point to instruct
JSCompiler to prune all files not reference by imports starting from
the named entry point.

In theory JSCompiler should be able to optimize-away all the
unreferenced stuff anyway, but since JSCompiler is not currently
optimizing well and since it will be easier to understand and
address the regression when there is less total output to examine,
this change is useful.

PiperOrigin-RevId: 107900067