Makes Arrays.function()'s inlinable to prepare for @nocollapse removal.
Arrays.impl.js currently issues some dynamic calls (calls on a Class
reference that was passed in as a parameter) to the
$isInstance/$isAssignable/$getClass static methods.

We want to remove the @nocollapse annotation (so we can get back a
bunch of optimization quality) but these dynamic calls in
Arrays.impl.js will get messed up.

So this patch moves all of Arrays.impl.js's dynamic method references
into very small top level functions that are ready for forced
inlining. Once they are inlined the @nocollapse removal will be safe.

PiperOrigin-RevId: 108054348