Removes dynamic this.constructor.$getClass() usage.
JSCompiler's CollapseProperties pass is a big win for optimized output
size but will break code that contains dynamic references to static fields.

This CL removes a dynamic reference (in Object.impl.js) to the $getClass()
static function and replaces that dynamism with a synthesized non-dynamic
m_getClass() implementation inside of every class. The new methods won't
add any output size once optimization is working again and the lack of
dynamism (in coordination with the upcoming J2clPass) will allow us to turn
CollapseProperties back on and get back a ton of size.

PiperOrigin-RevId: 108349697