stop requiring .native.js for native JsMethod.
Some native JsMethods refers to the Javascript built in function, and should not be expected to have any implementation provided by the user, for example, @JsMethod native boolean hasOwnProperty(). An ERR_NATIVE_JAVA_SOURCE_NO_MATCH error is only reported for the case in which a java type contains any non-JsMethod native methods, and does not have a matching native file.

PiperOrigin-RevId: 108564447