move array creation to a normalization pass.
We had a workaround for the JSCompiler bug b/24476009 that template inside a method body is unresolvable. The workaround replaced the template variable by its upper bound. But we output type cast in StatementSourceGeneartor for array creation statements, which is not normalized by this workaround, thus resulted in errors. Now moved the logic in StatementSourceGenerator to a normalization pass, thus the cast expression can be normalized. The normalization pass is not as clean as the String.format() in StatementSourceGenerator, but it is more maintainable.

PiperOrigin-RevId: 109747375