Adds Exception Java-JavaScript auto-conversion.
This it to make exceptions work well across Java/JavaScript boundary, following the recent similar design change in GWT. Also earlier, J2CL was not providing a way to catch exceptions thrown from JavaScript, this CL fixes that as well.

The essence of the change is, instead of throwing subclasses of Java Throwable class in generated throw statements, we now throw the backing JavaScript error object so the generated code looks like this:

  try {
    throw Exceptions.toJs(new SomeJavaException);
  } catch(e) {
    e = Exceptions.toJava(e);
    if (SomeJavaException.$isInstance(e)) {
      // catch block for SomeJavaException
    } else {
      throw Exceptions.toJs(e);
    }
  }

As the propagated thrown object is a real JavaScript error, it plays better with the browser dev tools (doesn't work well with custom error objects) and JavaScript callers (which become more essential due to JsInterop).

PiperOrigin-RevId: 110050780