Support for instanceof/cast check on JsFunction.
Like the way we do instanceof check on interface, we add a mark field "$is_mangled_type_name" on the instance of a JsFunctionImplementor, and use it to do instanceof check on JsFunctionImplementor.

As a result, instanceof check on Java instances follow the java semantics, and raw JS function instanceof JsFunctionImplementor returns false, raw JS function instnaceof JsFunctionInterface returns true.

PiperOrigin-RevId: 110471980