Fixes a bug in Js namespace processing on inner classes.
If you had an inner class 'Inner' in package 'com.google.j2cl'
inside of class 'Outer' then we would normally emit that as module
'com.google.j2cl.Outer$Inner'. If a custom JS namespace 'foo' was
provided we used to use it to replace the package and would emit
'foo.Outer$Inner' but that was wrong. The correct semantic is to
replace both the package AND the enclosing class names and emit
'foo.Inner'.

PiperOrigin-RevId: 110798877