Makes emul tests actually executed in JVM.
I recently noticed that our setup doesn't actually run any emul test in JVM although there is a target for that. This was because the GWTTestSuite replacement had an empty override for the method that suppose to record the tests to be executed. Removing the override fixes the issue and tests seems like executed now.

PiperOrigin-RevId: 110938106