fix bug for abstract inner class.
For an implicit qualifier of a method call, we match its declaring class with the nested classes stack from innermost to outermost. If the class isSubType() of the declaring class, then it is matched. But previously we only checked the super class, but missed super interfaces in isSubType(). It leads to wrong resolving in the case in which the declaring class is an interface.

PiperOrigin-RevId: 113766958