Always import 'Util' eagerly for Overlay class.
We did not import 'Util' for Overlay class, the reason was that Overlay class does not emit $getClass(), thus does not need Util. This is wrong. Util contains other utility functions besides $generateId( which is used in $getClass()), for example $makeLambdaFunction(). $makeLambdaFunction is emit at code generation stage, thus ImportGatheringVisitor missed importing it.

It seems redundant in some cases where Util is not really referenced, but is required. But we've already emit redundant imports in some cases, and this will not cause code regression after optimization. So I think it would be OK.

PiperOrigin-RevId: 113864715