Output JsDoc name for JsFunction interface and implementors as 'Function'.
Previously the JsDoc name for JsFunction interface/implementor is a union type of the type itself and 'Function'. It may break when integrating with native javascript code. We should allow the case that user specifies a javascript function that accepts a parameter of type 'Function' at Javascript side, and passes a JsFunction instance as an argument. But if the JsFunction instance is annotated as a union type, it cannot be passed to a slot that only accepts 'Function'.

Fixed JsDoc name for JsFunction type to be 'Function'. Then we need to specify @this type in JsFunction implementor's instance methods, otherwise it may cause type errors.

PiperOrigin-RevId: 114794469