Remove the $javaToString special case in toString trampoline.
This is not needed because toString already delegates to $javaToString for classes extending java.lang.Object. So we only need to special case Array.toString (special case for arrays is just for keeping the semantics same as java).

I also needed to update the tests as well because JsCompiler considers toString as side-effect free.

PiperOrigin-RevId: 116293681