Disable some JsCompiler flags for integration tests.
--closure_entry_point (and --only_closure_deps) both mask some code-size issues
and should thus be omitted for the purposes of knowing the size of the
transpiled output.

--remove_unused_constructor_properties is not safe to use for J2CL generate
code since it presumes that getters/setters are side-effect free (other than
the field that should be common between the two). This means that the
JsCompiler might ignore the clinit side-effect and prune the field and its
callers. This can end up producing incorrect code.

I know, I know, you're looking at this 175% increase in code size and thinking
this is crazy to commit. Thankfully almost all of this is going to be rectified
in some follow-up CLs shortly after.

PiperOrigin-RevId: 116988862