Adds --norewrite_polyfills to integration test size tracking.
Integration test size tracking aims to highlight size issues that
can be addressed by J2CL folk outside of JSCompiler or that are
expected to be addressed at some point inside of JSCompiler. The
polyfill size regression does not fit either of those categories
and so its size issue should not be highlighted.

PiperOrigin-RevId: 117157987