Makes TypeReference have the type of NativeFunction.
In java, type reference is not an expression. The TypeReference expression defined in J2CL is a reference to the constructor function
So the current claim that the type represented by TypeReference is the class it is referring to is conceptually incorrect.

I hit this because in one of my CLs; I had a TypeReference to long which compiler incorrectly tried to generate a boxing code because TypeReference expression claimed it is a type of long.

PiperOrigin-RevId: 117778177