Makes JSCompiler's implicit Object<V> -> Object<String, V> behavior explicit in J2CL.
JsInterop users are already depending upon this behavior and it's not something we
can roll back, but we need the support to be more complete (needs to work on type
aliases as well, which are going to stop working in JSCompiler as soon as the
goog.module() refactoring goes in).

Adding this support allowed JSCompiler to inspect types more accurately, and that
revealed bugs in InternalJsMap, which I fixed.

PiperOrigin-RevId: 118530138