Fix a bug when copying a UnionTypeDescriptor into a RegularTypeDescriptor, and add a repro case.
The problem was that previously certain accessors should never have been called on a UnionTypeDescriptor but with the ongoing refactoring to make RegularTypeDescriptor replace all other descriptors we now sometimes copy the entire contents of a descriptor out, slightly modify them, and put them into a RegularTypeDescriptor. This was triggering some of the UnsupportedOperations accessors in UnionTypeDescriptor.

This fixes the problem but also this won't be necessary at all after the next CL when UnionTypeDescriptor is deleted.

PiperOrigin-RevId: 119904606