Removing the spammiest part of the size report.
With this CL, if nothing has changed then you won't know if someone has actually run the report, but the upside is that when consecutive CLs don't make a change there will be no merge conflict.

Probably we can trust ourselves to run the report. If it turns out that we're not trustworthy, we can undo this change.

PiperOrigin-RevId: 120962551