Prevent inlining of Equality.$same by exporting it to global object that ensure Equality.$same is not inlined very early. This stabilizes the size tracking and avoid code size surprises in very small apps.
I tried adding a custom CodingConvention that prevents the inlining however that started causing other weirdnesses (like leaving Equality.same call with no args as it is never inlined and left us with big chunks of random code).

This is not the best solution but does the job for now. Created b/28676358 to track.

PiperOrigin-RevId: 121904710