Remove inner class $create helper.
Inner class create helper made more sense when we were considering publishing and using an ABI to j2cl code; without a public ABI there is no reason to keep them.

As an extra, these $create helpers were in lieu of new Innner() and in the cases like

  class A {
    class B {}
  }

  class A1 extends A {
    class B1 extends B {}
  }

resulted in polymorphic dispatch in contravention to Java semantics.

PiperOrigin-RevId: 121954283