Throw exception for null varargs.
Java and Javascript have slightly different semantics on varargs which introduces the need to wrap each spread operator with a $Util.$checkNotNul call.  In Java, if you call a vararg method/constructor with a single null parameter it will actually receive a null array rather than an array with a single null element.  In js, varargs aren't passed as an array so it is actually a single null parameter.  The spread of null "function(...null)" is a runtime exception in js, so we error out early with internal preconditions to warn users of this limitation.

For example this passes in java but fails in js.

  @JsMethod
  private static void method(int i, int...args) {
    assert args == null;
  }

  private static void testNullVararg() {
    method(1, null);
  }

Should throw an exception.

PiperOrigin-RevId: 122220707