Adds build macros for creating and for merging depgraphs (without verification).
Normally AJD verifies correctness in every merge but when synthesizing a
depgraph file that is a combination of a generated depgraph file for a ZIP and
extracted an depgraph from handrolled JS, the result only becomes valid after
the merge has completed (because of crosswise and circular dependencies).

While it's reasonable for the depgraphs of end user hand written JS to be
subject to constant verification, tools that synthesize JS and try to delegate
the depgraph construction process need to temporarily suspend verification to
be able to arrive at a final correct state. Thus the existence of this macro.

PiperOrigin-RevId: 124266827