Adds detection for incorrect flags in integration tests.
Earlier when a flag value is incorrect (e.g. missing quotes), it could cause an exception in uncompiled mode. When an exception is thrown from bootstrap, it will be silently ignored by jsunit test infra and as a result assertions wouldn't be enabled so all verifications become a no-op.
We actually shouldn't use assert statement for verification but that is a big change that I will not tackle.

This change ensures test fails by opening the alert dialog.

PiperOrigin-RevId: 126455753