Allows duplicate provide()s in AJD if they're in .JS.ZIP files.
This is needed since overlapping src globbing is relatively common
in Java, and when each java_library() becomes a Foo.js.zip file,
there will be duplicate provides.

The change also adds integration tests in J2CL that take advantage
of / prove the behavior.

The PMF Enhancer and JSCompiler tools that follow AJD in the toolchain
check for duplicate provides and *only* allow them if the content is
in a ZIP and the individual JS files are identical. Both these tools
also remove the duplicates if they are exact.

AJD would like to make the same ZIP allowance and enforce the same
exact content requirement but AJD does not know about the ZIP
contents, so only the ZIP allowance can be made here.

RELNOTES: none
PiperOrigin-RevId: 126460568