Speeds transpile time 26% by caching DescriptorFactory results.
I went looking for the cause of the slowdown in
ImportGatheringVisitor that Thomas identified, and found
the majority of it in repeated derivative TypeDescriptor
construction.

Since all derivative TypeDescriptors are created
in DescriptorFactory(s) it was pretty easy to add central
caching. This of course speeds up all derivative
TypeDescriptor creation, not just ImportGatheringVisitor.

I timed transpiling the standard library (containing 631
classes) 10 times.

Average time (measuring only J2CL) was 32238ms, 7537ms of
which was in ImportGatheringVisitor, so 23.4%.

After adding DescriptorFactory caching I measured 10 times
again.

Average time was 23852ms (26.0% overall reduction), 2359ms of
which was in ImportGatheringVisitor, so now 9.8% of total.

PiperOrigin-RevId: 126758829