Consistently put "__" in mangled names, even when there are no parameters.
This is to make sure that even there are no parameters for the method, we still generate a mangled name similar to case with parameters and it will be still obvious that the name is internal (e.g. $create vs $create__, m_get vs m_get__).

Public:
Update Xplat pass to match J2CL mangling change.
PiperOrigin-RevId: 126967085