Fix unnecessary boolean expression
This change fixes boolean expressions like `a && a` and `a || a` where both
arguments are the same, by replacing the entire expression with `a`.

The clean up is being done in preparation for enabling a build error.

*Note to reviewers* the intent may have been to use two different expressions
(e.g. `a && b`). If this isn't the best fix, please suggest an edit in
critique.

More information:
https://docs.google.com/a/google.com/document/d/1FN2oqU8QmJ2aVSx969rORp-nWcfgdVjqZ6KOFyxYcHY/edit?usp=sharing

Tested:
    TAP --sample for global presubmit queue
    http://test/OCL:127546952:BASE:127579671:1468628558893:f6a198c2
PiperOrigin-RevId: 127597244