Allow Java annotation processors to produce native.js files.
Currently annotation processors in blaze can only write resources to the jar that contains class files. We may want to look in working with the blaze team to change that.

This CL introduces a genrule that takes the resource jar and produces a zip file from it containing potential native.js files produced in an APT.
This is needed for the j2cl Polymer integration I am working on.

PiperOrigin-RevId: 127993881