Removes InsertTypeAnnotationOnGenericReturnTypes.
It was originally put in place to work around a behavior in JSCompiler
where type inference would specialize a generic type argument to a
non-null value and then cause a type check error when JSCompiler
was unable to covariantly resolve compatibility between Foo<Bar> and
Foo<!Bar>.

By inserting Foo<?> casts we prevented JSCompiler from performing type
inference that would break the code.

But this caused problems with J2CL + AJD since the these inserted
casts could be inserted in the middle of complex expressions and so
force importing types that were not imported in the original Java
source. These extra imports violated strict dependency checking.

Luckily, JSCompiler now ignores nullability mismatches in J2CL output,
so there is no longer a need to insert these casts.

Removing the casts revealed some previously-hidden type mismatches in
InternalHashCodeMap/InternalJsMap/InternalStringMap for which the fixes
are included in this CL.

Change-Id: I7e437cbac90a1fced21dd4fed937bd7ff89a7795
PiperOrigin-RevId: 128418577