Create a local alias to the class defined in the module for native.js.
Often when implementing native methods in a <Class>.native.js file, the hand written code need to reference to the name of the class being defined in the module.

In most cases the name of the class is just the Java class name; however, in some cases the compiler might choose a different name due to conflicts. These conflicts will become more prevalent when j2cl stops aliasing externs.

This cl creates an alias " __class" to the class defined in the current model for use in .native.js files.

PiperOrigin-RevId: 128848578