Automated g4 rollback of changelist 129052387.
*** Reason for rollback ***

Breaks xplat.

*** Original change description ***

Stop aliasing externs (take 2).

Native JsTypes in JsPackage.GLOBAL namespace are considered externs, so they are not goog.required, instead they are directly aliased by a let definition.

This approach does not work well if the extern was a @typedef and not a @constructor.

This CL stops aliasing externs and instead locally renames definitions that would be conflicting with the externs.

***

PiperOrigin-RevId: 129116554