Automated g4 rollback of changelist 128761287.
*** Reason for rollback ***

Doc's JS isn't compatible. Temporarily rolling back till we make a final decision.

*** Original change description ***

Creates Java's implicit equals and hashCode methods in top-most interfaces.

This gives JS users the same experience as Java users in that they can treat variables (which are typed as some interface) as being able to dispatch the mangled forms of equals() and hashCode(), without the user having to know about the existence of Objects.js' trampolining.

***

PiperOrigin-RevId: 129482576