Makes replace_all.py build.log generation less finicky.
Instead of removing files from blaze output to force recompile, we now just let blaze recompile only the binaries that need recompiling and only copy/delete the logs for the readable targets that were recompiled.

This speeds up the process slightly in general and substantially when only readable a few examples changed.

PiperOrigin-RevId: 130054080