Remove usage of System.exit from the compiler.
Instead we throw a specific error which is caught at the top level and then
causes a System.exit.

Our upcoming worker uses a different entrypoint and thus can catch this Error.
This will also allow the worker to distinguish between a graceful exit and an
actual error in the compiler.

PiperOrigin-RevId: 130235322