Avoid some TemplateTypeMap duplication during subtype checking.
This reduces 'checkTypes' time in the J2CL 'timing' readable example
(which uses Guava) from 24.5 seconds to 1.7 seconds.

For future reference, note that when performing subtype analysis
JSCompiler sometimes incidentally calls
TemplateTypeMap.getResolvedTemplateType(). With that out of the way we
continue:

It just so happens that the TemplateTypeMap datastructure allows its
contained list of template keys and the list of template values to
have different lengths. When responding to getResolvedTemplateType()
requests it wants to correct this length mismatch (by duplicating the
map and filling in the missing template values slots with UNKNOWN_TYPE)
so that arbitrary index accesses will be safe.

But the act of duplicating a TemplateTypeMap and inserting UNKNOWN_TYPE
triggers new (and also expensive) type comparisons.

If the types being dealt with are nested templatized types, the result is
a super-linear explosion of back and forth cross talk between the acts of
subtype checking and TemplateTypeMap duplicating + UNKNOWN_TYPE padding.

The simple fix is to make arbitrary index access in
getResolvedTemplateType() safe not by duplicating and padding with
UNKNOWN_TYPE but to instead directly return UNKNOWN_TYPE if the index is
past the bounds of the template values list.

PiperOrigin-RevId: 131610944