Allow use of namespace="window" to be similar to JsPackage.GLOBAL.
In GWT, JsPackage.GLOBAL refers to the top window (as opposed as the iframe which in some cases is wanted) and accesses are prepended with "$wnd.". This CL makes the namespace "window" to refer to the iframe top scope instead.

In J2cl this CL makes namespace "window" to behave exactly as JsPackage.GLOBAL for compatibiliy with GWT semantics.

PiperOrigin-RevId: 131730481