Speeds up full opt J2CL builds and removes the temporary limitJ2clOptimization flag.
Lowers a fully optimized PhotosUI (with integrated J2CL) build from:
532 passes, http://sponge/79d59bb8-3784-473a-b506-6762e0c0c32f
to
357 passes, http://sponge/6338ee2a-cfa0-451f-821e-35044b235a30

Lowers a fully optimized BigTop build (yes i turned on all opts) from:
402 passes, http://sponge/e76f7b81-decd-418f-b59e-3a8232c93151
to
373 passes, http://sponge/2aa3fa4a-9630-4511-b482-b3eeff135057

The changes are:
- force the 3 J2CL opt passes to stay on permanently (unlike other PhaseOptimizer managed passes)
- remove J2CL opt passes from the "standard optimizations" loop but keep them in the "main optimizations" loop
- remove the limitJ2clOptimization flag

The reason these changes have such an impact is because (unlike other passes) the J2CL opt passes have a tendency to toggle randomly between making changes and not making changes. The consequence of this is that PhaseOptimizer would turn them off pretty early and so delay their cascade of side effect changes to much later full optimize loop resets. This resulted in an expensive cascade of multiple full optimize loop resets.

Forcing them to stay on causes their changes to occur earlier, thus cascaded side effect changes occur earlier (much of the cost of that amortized into the cost of other pass runs that were going to occur anyway) and finally leading to less late work being discovered in new optimize loop resets.

But forcing these passes to stay on causes a lot of overhead. This overhead was mostly visible in the "standard optimizations" loop (which happens to repeat very many times). Removing the J2CL passes from the "standard optimizations" loop saved this time, but also it made sense to do so since the passes are not "code removing" passes and probably shouldn't have been in the "standard optimizations" loop to begin with.

Public:
Speeds up full opt J2CL builds and removes the temporary limitJ2clOptimization flag.
PiperOrigin-RevId: 132588230