Simplify code output for array of j.l.Object and native type.
Instead of using Arrays.$create, now the simpler form of new Array() is output.
This prevents extra metadata to be added to this forms of array which has no use at the runtime and polute the array object.

Note that unlike GWT, multi dimension arrays of native types still use the old form.

PiperOrigin-RevId: 132750717