Use sharding for JRE tests.
Now that jsunit_test does understand zips our tests are no longer one gigantic
tests that can not be sharded.

Sharding in JS works per suite (per JUnit test) and thus we can run lots of
them in paralell. On my machine this is roughly a 2-4x speed up for
//third_party/java_src/j2cl/jre/javatests:Emul_compiled

PiperOrigin-RevId: 134584013