Record a file-end source mapping so that enough ';' groups are created.
The JSCompiler SourceMap file construction library needs to emit a ';' group
entry for every line in the JS output. But it only knows about what lines
exist based on the mappings that are provided.

If there are lines at the end of a file that are not mapped, it will not
create ';' group entries for them since it does not know the file length.
The lack of a function for informing it of file length is probably a bug in
the library, I will report that bug separately.

That is why we have started recording a fake mapping at the very end of the
file.

PiperOrigin-RevId: 134590707